Make ExampleFromFile parse an invalid example so that later it would be caught in example validation. Keep the behavior of loading externalised examples the same for tests and stub
