+++ core/src/main/kotlin/io/specmatic/core/Feature.kt
@@ -1673,13 +1673,22 @@ data class Feature(
                 acc + loadExternalisedJSONExamples(item)
             }
 
-        return examplesInSubdirectories + files.filterNot { it.isDirectory }.map { ExampleFromFile(it) }.mapNotNull { exampleFromFile ->
+        logger.log("Loading externalised examples in ${testsDirectory.path}: ")
+        return examplesInSubdirectories + files.asSequence().filterNot {
+            it.isDirectory
+        }.map {
+            val exampleFromFile = ExampleFromFile(it)
+            if(exampleFromFile.isInvalid()) {
+                throw ContractException("Error loading example from file '${it.name}' as it is in invalid format. Please fix the example format to load this example.")
+            }
+            exampleFromFile
+        }.mapNotNull { exampleFromFile ->
             try {
                 with(exampleFromFile) {
                     OpenApiSpecification.OperationIdentifier(
-                        requestMethod,
-                        requestPath,
-                        responseStatus,
+                        requestMethod.orEmpty(),
+                        requestPath.orEmpty(),
+                        responseStatus ?: 0,
                         exampleFromFile.headers.filter { it.key.lowercase() == "content-type" }.values.firstOrNull(),
                         exampleFromFile.responseHeaders?.let { it.jsonObject.filter { it.key.lowercase() == "content-type" }.values.firstOrNull()?.toStringLiteral() }
                     ) to exampleFromFile.toRow(specmaticConfig)
