+++ core/src/main/kotlin/io/specmatic/core/examples/server/ExamplePageRequest.kt
@@ -1,0 +1,6 @@
+package io.specmatic.core.examples.server
+
+data class ExamplePageRequest(
+    val contractFile: String,
+    val hostPort: String
+)
