+++ core/src/main/kotlin/io/specmatic/core/examples/server/ExampleTestRequest.kt
@@ -1,0 +1,5 @@
+package io.specmatic.core.examples.server
+
+data class ExampleTestRequest(
+    val exampleFile: String
+)
