+++ core/src/main/kotlin/io/specmatic/core/examples/server/GenerateExample.kt
@@ -1,0 +1,6 @@
+package io.specmatic.core.examples.server
+
+data class GenerateExample(
+    val exampleFilePath: String,
+    val created: Boolean
+)
