+++ core/src/main/kotlin/io/specmatic/core/examples/server/SaveExampleRequest.kt
@@ -1,0 +1,6 @@
+package io.specmatic.core.examples.server
+
+data class SaveExampleRequest(
+    val exampleFile: String,
+    val exampleContent: String
+)
