+++ core/src/main/kotlin/io/specmatic/core/examples/server/ValidateExampleRequest.kt
@@ -1,0 +1,5 @@
+package io.specmatic.core.examples.server
+
+data class ValidateExampleRequest(
+    val exampleFile: String
+)
