+++ core/src/main/kotlin/io/specmatic/core/examples/server/ValidateExampleResponse.kt
@@ -1,0 +1,8 @@
+package io.specmatic.core.examples.server
+
+data class ValidateExampleResponse(
+    val absPath: String,
+    val errorMessage: String? = null,
+    val errorList: List<Map<String, String>> = emptyList(),
+    val isPartialFailure: Boolean = false
+)
