+++ core/src/main/kotlin/io/specmatic/core/examples/server/ValidateExampleResponseMap.kt
@@ -1,0 +1,7 @@
+package io.specmatic.core.examples.server
+
+data class ValidateExampleResponseMap(
+    val absPath: String,
+    val error: List<Map<String, Any?>> = emptyList(),
+    val isPartialFailure: Boolean = false
+)
